+++ fastchart_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: 8214d78acfcdcec7bd92398f069ae29dbfb2dbc2 */
+ * Stub hash: bfa09db1fcbcf923d01d3b67b4dae4b54132cab9 */
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_class_FastChart_Chart___construct, 0, 0, 0)
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, width, IS_LONG, 1, "null")
@@ -886,6 +886,24 @@ static zend_class_entry *register_class_FastChart_Chart(void)
 	zend_string *const_STYLE_I_CAP_name = zend_string_init_interned("STYLE_I_CAP", sizeof("STYLE_I_CAP") - 1, 1);
 	zend_declare_typed_class_constant(class_entry, const_STYLE_I_CAP_name, &const_STYLE_I_CAP_value, ZEND_ACC_PUBLIC, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
 	zend_string_release(const_STYLE_I_CAP_name);
+
+	zval const_STYLE_HOLLOW_value;
+	ZVAL_LONG(&const_STYLE_HOLLOW_value, 4);
+	zend_string *const_STYLE_HOLLOW_name = zend_string_init_interned("STYLE_HOLLOW", sizeof("STYLE_HOLLOW") - 1, 1);
+	zend_declare_typed_class_constant(class_entry, const_STYLE_HOLLOW_name, &const_STYLE_HOLLOW_value, ZEND_ACC_PUBLIC, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
+	zend_string_release(const_STYLE_HOLLOW_name);
+
+	zval const_STYLE_VOLUME_value;
+	ZVAL_LONG(&const_STYLE_VOLUME_value, 5);
+	zend_string *const_STYLE_VOLUME_name = zend_string_init_interned("STYLE_VOLUME", sizeof("STYLE_VOLUME") - 1, 1);
+	zend_declare_typed_class_constant(class_entry, const_STYLE_VOLUME_name, &const_STYLE_VOLUME_value, ZEND_ACC_PUBLIC, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
+	zend_string_release(const_STYLE_VOLUME_name);
+
+	zval const_STYLE_VECTOR_value;
+	ZVAL_LONG(&const_STYLE_VECTOR_value, 6);
+	zend_string *const_STYLE_VECTOR_name = zend_string_init_interned("STYLE_VECTOR", sizeof("STYLE_VECTOR") - 1, 1);
+	zend_declare_typed_class_constant(class_entry, const_STYLE_VECTOR_name, &const_STYLE_VECTOR_value, ZEND_ACC_PUBLIC, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
+	zend_string_release(const_STYLE_VECTOR_name);
 
 	zval const_BORDER_NONE_value;
 	ZVAL_LONG(&const_BORDER_NONE_value, 0);